Pass --disable-factory only to gnome-terminal version <3.8
Fixes "Silently fails to open terminal with gnome-terminal >= 3.8" problem.
This option was required for gnome-terminal <3.6 to keep environment vars.
It was removed in 3.8 and raises error with later versions.
Fortunately it is not needed in later versions: environment is always kept.
There is no obvious stable & documented way to determine version
of gnome-terminal. So just check the presence of a file that was used by
versions 2.24-3.6. This slightly breaks old (pre Sep2008)
gnome-terminal <2.24 by not passing environment vars to the launched script