Fix rpdb2 when COLORTERM environment variable is set to 'yes'
On some systems, the COLORTERM environment variable is set to 'yes'.  When
this happens, rpdb2 attempts to use yes as the terminal application.  Since
this is a valid binary that is not a terminal application, this will cause
rpdb2 to fail to open a terminal.  Avoid this by excluding 'yes' as a
possible value when checking COLORTERM.