Do not print if no file was provided.
The current Python 3.13.2 documentation states [https://docs.python.org/3/library/sys.html#sys.__stderr__]:
"Note: Under some conditions stdin, stdout and stderr as well as the original values __stdin__, __stdout__ and __stderr__ can be None. It is usually the case for Windows GUI apps that aren’t connected to a console and Python apps started with pythonw."

Using an integrated Python interpreter, I encountered sys.__stderr__ to be None - which prevented me from using rpdb as it crashed when printing debug information in print_debug().

In addition I like using a None file-handle to disable output to it. I find this to be pythonic.